evm: split testing scripts
