+++ evm/ts/test/02_send_receive.ts
@@ -40,7 +40,7 @@ import {
   findRedeemEventInLogs,
 } from "./helpers/utils";
 
-describe("Circle Integration Test", () => {
+describe("Circle Integration Send and Receive", () => {
   // ethereum wallet, CircleIntegration contract and USDC contract
   const ethProvider = new ethers.providers.StaticJsonRpcProvider(ETH_LOCALHOST);
   const ethWallet = new ethers.Wallet(WALLET_PRIVATE_KEY, ethProvider);
@@ -82,237 +82,6 @@ describe("Circle Integration Test", () => {
     AVAX_WORMHOLE_ADDRESS,
     avaxWallet
   );
-
-  describe("Registrations", () => {
-    // produces governance VAAs for CircleAttestation contract
-    const governance = new CircleGovernanceEmitter();
-
-    describe("Ethereum Goerli Testnet", () => {
-      it("Should Register Foreign Circle Integration", async () => {
-        const timestamp = getTimeNow();
-        const chainId = await ethCircleIntegration.chainId();
-        const emitterChain = await avaxCircleIntegration.chainId();
-        const emitterAddress = Buffer.from(
-          tryNativeToUint8Array(avaxCircleIntegration.address, "avalanche")
-        );
-        const domain = await avaxCircleIntegration.localDomain();
-
-        // create unsigned registerEmitterAndDomain governance message
-        const published =
-          governance.publishCircleIntegrationRegisterEmitterAndDomain(
-            timestamp,
-            chainId,
-            emitterChain,
-            emitterAddress,
-            domain
-          );
-
-        // sign governance message with guardian key
-        const signedMessage = guardians.addSignatures(published, [0]);
-
-        // register the emitter and domain
-        const receipt = await ethCircleIntegration
-          .registerEmitterAndDomain(signedMessage)
-          .then((tx) => tx.wait())
-          .catch((msg) => {
-            // should not happen
-            console.log(msg);
-            return null;
-          });
-        expect(receipt).is.not.null;
-
-        // check contract state to verify the registration
-        const registeredEmitter = await ethCircleIntegration
-          .getRegisteredEmitter(emitterChain)
-          .then((bytes) => Buffer.from(ethers.utils.arrayify(bytes)));
-        expect(Buffer.compare(registeredEmitter, emitterAddress)).to.equal(0);
-      });
-
-      it("Should Register Accepted Token", async () => {
-        const timestamp = getTimeNow();
-        const chainId = await ethCircleIntegration.chainId();
-
-        // create unsigned registerAcceptedToken governance message
-        const published =
-          governance.publishCircleIntegrationRegisterAcceptedToken(
-            timestamp,
-            chainId,
-            ETH_USDC_TOKEN_ADDRESS
-          );
-
-        // sign governance message with guardian key
-        const signedMessage = guardians.addSignatures(published, [0]);
-
-        // register the token
-        const receipt = await ethCircleIntegration
-          .registerAcceptedToken(signedMessage)
-          .then((tx) => tx.wait())
-          .catch((msg) => {
-            // should not happen
-            console.log(msg);
-            return null;
-          });
-        expect(receipt).is.not.null;
-
-        // check contract state to verify the registration
-        const accepted = await ethCircleIntegration.isAcceptedToken(
-          ETH_USDC_TOKEN_ADDRESS
-        );
-        expect(accepted).is.true;
-      });
-
-      it("Should Register Target Chain Token", async () => {
-        const timestamp = getTimeNow();
-        const chainId = await ethCircleIntegration.chainId();
-        const targetChain = await avaxCircleIntegration.chainId();
-        const targetToken = Buffer.from(
-          tryNativeToUint8Array(AVAX_USDC_TOKEN_ADDRESS, "avalanche")
-        );
-
-        // create unsigned registerTargetChainToken governance message
-        const published =
-          governance.publishCircleIntegrationRegisterTargetChainToken(
-            timestamp,
-            chainId,
-            ETH_USDC_TOKEN_ADDRESS,
-            targetChain,
-            targetToken
-          );
-
-        // sign governance message with guardian key
-        const signedMessage = guardians.addSignatures(published, [0]);
-
-        // register the target token
-        const receipt = await ethCircleIntegration
-          .registerTargetChainToken(signedMessage)
-          .then((tx) => tx.wait())
-          .catch((msg) => {
-            // should not happen
-            console.log(msg);
-            return null;
-          });
-        expect(receipt).is.not.null;
-
-        // check contract state to verify the registration
-        const registeredTargetToken = await ethCircleIntegration
-          .targetAcceptedToken(ETH_USDC_TOKEN_ADDRESS, targetChain)
-          .then((bytes) => Buffer.from(ethers.utils.arrayify(bytes)));
-        expect(Buffer.compare(registeredTargetToken, targetToken)).to.equal(0);
-      });
-    });
-
-    describe("Avalanche Fuji Testnet", () => {
-      it("Should Register Foreign Circle Integration", async () => {
-        const timestamp = getTimeNow();
-        const chainId = await avaxCircleIntegration.chainId();
-        const emitterChain = await ethCircleIntegration.chainId();
-        const emitterAddress = Buffer.from(
-          tryNativeToUint8Array(ethCircleIntegration.address, "avalanche")
-        );
-        const domain = await ethCircleIntegration.localDomain();
-
-        // create unsigned registerEmitterAndDomain governance message
-        const published =
-          governance.publishCircleIntegrationRegisterEmitterAndDomain(
-            timestamp,
-            chainId,
-            emitterChain,
-            emitterAddress,
-            domain
-          );
-        const signedMessage = guardians.addSignatures(published, [0]);
-
-        // sign governance message with guardian key
-        const receipt = await avaxCircleIntegration
-          .registerEmitterAndDomain(signedMessage)
-          .then((tx) => tx.wait())
-          .catch((msg) => {
-            // should not happen
-            console.log(msg);
-            return null;
-          });
-        expect(receipt).is.not.null;
-
-        // check contract state to verify the registration
-        const registeredEmitter = await avaxCircleIntegration
-          .getRegisteredEmitter(emitterChain)
-          .then((bytes) => Buffer.from(ethers.utils.arrayify(bytes)));
-        expect(Buffer.compare(registeredEmitter, emitterAddress)).to.equal(0);
-      });
-
-      it("Should Register Accepted Token", async () => {
-        const timestamp = getTimeNow();
-        const chainId = await avaxCircleIntegration.chainId();
-
-        // create unsigned registerAcceptedToken governance message
-        const published =
-          governance.publishCircleIntegrationRegisterAcceptedToken(
-            timestamp,
-            chainId,
-            AVAX_USDC_TOKEN_ADDRESS
-          );
-
-        // sign governance message with guardian key
-        const signedMessage = guardians.addSignatures(published, [0]);
-
-        // register the token
-        const receipt = await avaxCircleIntegration
-          .registerAcceptedToken(signedMessage)
-          .then((tx) => tx.wait())
-          .catch((msg) => {
-            // should not happen
-            console.log(msg);
-            return null;
-          });
-        expect(receipt).is.not.null;
-
-        // check contract state to verify the registration
-        const accepted = await avaxCircleIntegration.isAcceptedToken(
-          AVAX_USDC_TOKEN_ADDRESS
-        );
-        expect(accepted).is.true;
-      });
-
-      it("Should Register Target Chain Token", async () => {
-        const timestamp = getTimeNow();
-        const chainId = await avaxCircleIntegration.chainId();
-        const targetChain = await ethCircleIntegration.chainId();
-        const targetToken = Buffer.from(
-          tryNativeToUint8Array(ETH_USDC_TOKEN_ADDRESS, "avalanche")
-        );
-
-        // create unsigned registerTargetChainToken governance message
-        const published =
-          governance.publishCircleIntegrationRegisterTargetChainToken(
-            timestamp,
-            chainId,
-            AVAX_USDC_TOKEN_ADDRESS,
-            targetChain,
-            targetToken
-          );
-
-        // sign governance message with guardian key
-        const signedMessage = guardians.addSignatures(published, [0]);
-
-        // register the target token
-        const receipt = await avaxCircleIntegration
-          .registerTargetChainToken(signedMessage)
-          .then((tx) => tx.wait())
-          .catch((msg) => {
-            // should not happen
-            console.log(msg);
-            return null;
-          });
-        expect(receipt).is.not.null;
-
-        /// check contract state to verify the registration
-        const registeredTargetToken = await avaxCircleIntegration
-          .targetAcceptedToken(AVAX_USDC_TOKEN_ADDRESS, targetChain)
-          .then((bytes) => Buffer.from(ethers.utils.arrayify(bytes)));
-        expect(Buffer.compare(registeredTargetToken, targetToken)).to.equal(0);
-      });
-    });
-  });
 
   describe("Transfer With Payload Logic", () => {
     const amountFromEth = ethers.BigNumber.from("69");
